Linker fix - need to move init_resources along.
